+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.17.44
 ### is.codion
 - ServiceConfigurationError caught, cause unwrapped and rethrown.
+### is codion.common.core
+- Text.SpaceAwareCollator bug fixed, now treats null values as an empty string.
 ### is codion.common.db
 - Database.databaseException() renamed exception().
 ### is.codion.swing.common.model